create server/utils/cache.js with maintainCache() functionality. Remove existing definitions of same function. Load in server.js and pass in ctx object.
